+++ alacritty/src/config/mod.rs
@@ -204,6 +204,14 @@ fn print_deprecation_warnings(config: &Config) {
              lines"
         );
     }
+
+    if config.scrolling.auto_scroll.is_some() {
+        warn!(
+            target: LOG_TARGET_CONFIG,
+            "Config scrolling.auto_scroll has been removed and is now always disabled, it can be \
+             safely removed from the config"
+        );
+    }
 }
 
 #[cfg(test)]